+++ spec/enotas_api/v1/baixar_pdf_nfs_id_externo.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe EnotasApi::V1::BaixarPdfNfsIdExterno do
+  let(:empresa_id) { '123' }
+  let(:id_externo) { '123' }
+  let(:instance) { described_class.new(empresa_id, id_externo) }
+
+  it 'have expected uri' do
+    expect(instance.uri).to eq("/v1/empresas/#{empresa_id}/nfes/porIdExterno/#{id_externo}/pdf")
+  end
+
+  it 'have expected result' do
+    expect(instance.result_object).to eq(EnotasApi::RawResult)
+  end
+end