+++ spec/enotas_api/v1/baixar_xml_nfs_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe EnotasApi::V1::BaixarXmlNfs do
+  let(:empresa_id) { '123' }
+  let(:nfs_id) { '123' }
+  let(:instance) { described_class.new(empresa_id, nfs_id) }
+
+  it 'have expected uri' do
+    expect(instance.uri).to eq("/v1/empresas/#{empresa_id}/nfes/#{nfs_id}/xml")
+  end
+
+  it 'have expected result' do
+    expect(instance.result_object).to eq(EnotasApi::RawResult)
+  end
+end